Shorter readme intro (#3)
My info is of course on GitHub, and if I publish a wheel on PyPI,
the maintainer metadata should be shown. The important things, and
chief goals of the intro, are just to make clear that the upstream
developers aren't responsible for this fork and to make clear that
I shouldn't be credited for their code.